requirements.txt
requests
python-dotenv
prefect
mysql-connector-python
tabulate